+++ node/src/tests/runtime_api_mock.rs
@@ -100,6 +100,10 @@ sp_api::mock_impl_runtime_apis! {
 
 			}
 		}
+
+		fn initialized() -> bool {
+			true
+		}
 	}
 }