+++ primitives/native-token-management/src/tests/runtime_api_mock.rs
@@ -31,6 +31,10 @@ sp_api::mock_impl_runtime_apis! {
 			MainChainScripts::default()
 		}
 
+		fn initialized() -> bool {
+			true
+		}
+
 	}
 }